requirements.txt
Flask==3.0.3
flask-cors==4.0.1

flask-pymongo==2.3.0
pymongo==4.6.1

PyJWT==2.9.0
cryptography==42.0.8
python-dotenv==1.0.1

pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3